requirements.txt
jinja2
django
djangorestframework
markdown
django-filter
dj_database_url
psycopg2
pillow